+++ lib/pkgxx/tty.cxx
@@ -6,6 +6,7 @@
 #include <system_error>
 #include <unistd.h>
 
+#include "environment.hxx"
 #include "tty.hxx"
 
 namespace pkgxx {
@@ -16,14 +17,6 @@ namespace pkgxx {
 #else
         return false;
 #endif
-    }
-
-    ttystream::ttystream(int fd, bool owned)
-        : fdstream(fd, owned) {
-
-        if (!cisatty(fd)) {
-            throw not_a_tty(fd);
-        }
     }
 
     std::optional<dimension<std::size_t>>
@@ -37,10 +30,7 @@ namespace pkgxx {
         if (ioctl(*fd(), TIOCGWINSZ, &ws) != 0) {
             throw std::system_error(errno, std::generic_category(), "ioctl(TIOCGWINSZ)");
         }
-        return dimension<std::size_t> {
-            .width  = ws.ws_col,
-            .height = ws.ws_row
-        };
+        return dimension<std::size_t> { ws.ws_col, ws.ws_row };
 
 #else
         // Dangit, this platform doesn't have TIOCGWINSZ. It's probably
@@ -49,12 +39,18 @@ namespace pkgxx {
         // We could open the controlling terminal device, temporarily turn
         // off ICANON and ECHO, move the cursor to the rightmost column,
         // request the cursor position to tty, read a response from tty,
-        // and then parse it. But this complicates matter a lot, like damn
-        // fucking a LOT. We can't bear with the complexity in this already
-        // suboptimal case. No operating systems relevant today lack
-        // TIOCGWINSZ after all.
+        // and then parse it. But that complicates matters a lot, like damn
+        // fucking a LOT. We can't bear with the complexity it induces in
+        // this already suboptimal case. No operating systems relevant
+        // today lack TIOCGWINSZ after all.
         return std::nullopt;
 #endif
+    }
+
+    bool
+    ttystream::default_use_colour() {
+        // See https://no-color.org/
+        return !cgetenv("NO_COLOR").has_value();
     }
 
     namespace tty {
@@ -62,10 +58,10 @@ namespace pkgxx {
             ttystream&
             operator<< (ttystream& tty, move_to const& m) {
                 if (m.y) {
-                    tty << "\e[" << *m.y + 1 << ';' << m.x + 1 << 'H';
+                    tty << "\x1B[" << *m.y + 1 << ';' << m.x + 1 << 'H';
                 }
                 else if (m.x > 0) {
-                    tty << "\e[" << m.x + 1 << 'F';
+                    tty << "\x1B[" << m.x + 1 << 'F';
                 }
                 else {
                     tty << '\r';
@@ -76,8 +72,108 @@ namespace pkgxx {
 
         ttystream&
         erase_line_from_cursor(ttystream& tty) {
-            tty << "\e[K";
+            tty << "\x1B[K";
             return tty;
         }
+
+        style&
+        style::operator+= (style const& rhs) noexcept {
+            if (!foreground && rhs.foreground) foreground = rhs.foreground;
+            if (!background && rhs.background) background = rhs.background;
+            if (!boldness   && rhs.boldness  ) boldness   = rhs.boldness;
+            if (!font       && rhs.font      ) font       = rhs.font;
+            if (!underline  && rhs.underline ) underline  = rhs.underline;
+            return *this;
+        }
+
+        ttystream&
+        operator<< (ttystream& tty, style const& sty) {
+            if (!tty.use_colour()) {
+                return tty;
+            }
+
+            tty << "\x1B[0";
+            if (sty.foreground) {
+                auto const& [i, c] = *sty.foreground;
+                tty << ';'
+                    << 30 + static_cast<int>(i) + static_cast<int>(c);
+            }
+            if (sty.background) {
+                auto const& [i, c] = *sty.background;
+                tty << ';'
+                    << 40 + static_cast<int>(i) + static_cast<int>(c);
+            }
+            if (sty.boldness) {
+                tty << ';'
+                    << static_cast<int>(*sty.boldness);
+            }
+            if (sty.font) {
+                tty << ';'
+                    << static_cast<int>(*sty.font);
+            }
+            if (sty.underline) {
+                tty << ';'
+                    << static_cast<int>(*sty.underline);
+            }
+            tty << 'm';
+            return tty;
+        }
+
+        style
+        dull_colour(detail::colour const c) {
+            return style {
+                std::make_optional(
+                    std::make_pair(
+                        detail::intensity::dull, c)),
+                {}, {}, {}, {}
+            };
+        }
+
+        style
+        colour(detail::colour const c) {
+            return style {
+                std::make_optional(
+                    std::make_pair(
+                        detail::intensity::vivid, c)),
+                {}, {}, {}, {}
+            };
+        }
+
+        style
+        dull_bg_colour(detail::colour const c) {
+            return style {
+                {},
+                std::make_optional(
+                    std::make_pair(
+                        detail::intensity::dull, c)),
+                {}, {}, {}
+            };
+        }
+
+        style
+        bg_colour(detail::colour const c) {
+            return style {
+                {},
+                std::make_pair(
+                    detail::intensity::vivid, c),
+                {}, {}, {}
+            };
+        }
+
+        style const bold = style {
+            {}, {}, detail::boldness::bold, {}, {}
+        };
+
+        style const faint = style {
+            {}, {}, detail::boldness::faint, {}, {}
+        };
+
+        style const italicised = style {
+            {}, {}, {}, detail::font::italics, {}
+        };
+
+        style const underlined = style {
+            {}, {}, {}, {}, detail::underline::single
+        };
     }
 }